basic impl of expert and admin bots

+++ api/src/main/scala/com/github/ppotseluev/itdesk/api/telegram/package.scala
@@ -2,6 +2,6 @@ package com.github.ppotseluev.itdesk.api
 
 package object telegram {
   type WebhookSecret = String
-  type UserId = Int
+  type UserId = Long
   type BotToken = String
 }
